Add Red and Blue Activator Area
You can use the Q and E keys to toggle the behavior of red and blue obstacles. Next, I'll work on an indication for the user, such as glowing the red and blue UI elements, along with a scaling letter animation of 'Q' and 'E' near the player to let user know when they can use it.