fix: shaderPrecisionModel color washout
I have fixed the issue that made the colors washed out. It was just one value, shaderPrecisionModel, that needed to be changed from 1 to 0. It required a ton of debugging. I have tested it multiple times on my PC and the colors are fine. Let me know if the colors are fine in your project as well.